+++ src/core/HierarchicalStorage.h
@@ -50,7 +50,7 @@ public:
   virtual void __fastcall WriteInt64(const UnicodeString Name, __int64 Value) = 0;
   virtual void __fastcall WriteDateTime(const UnicodeString Name, TDateTime Value) = 0;
   virtual void __fastcall WriteFloat(const UnicodeString Name, double Value) = 0;
-  virtual void __fastcall WriteBinaryData(const UnicodeString Name, const void * Buffer, int Size) = 0;
+  virtual void __fastcall WriteBinaryData(const UnicodeString Name, const void * Buffer, size_t Size) = 0;
 
   virtual void __fastcall WriteString(const UnicodeString Name, const UnicodeString Value);
   void __fastcall WriteBinaryData(const UnicodeString Name, const RawByteString Value);
@@ -119,7 +119,7 @@ public:
   virtual void __fastcall WriteDateTime(const UnicodeString Name, TDateTime Value);
   virtual void __fastcall WriteFloat(const UnicodeString Name, double Value);
   virtual void __fastcall WriteStringRaw(const UnicodeString Name, const UnicodeString Value);
-  virtual void __fastcall WriteBinaryData(const UnicodeString Name, const void * Buffer, int Size);
+  virtual void __fastcall WriteBinaryData(const UnicodeString Name, const void * Buffer, size_t Size);
 
   virtual void __fastcall GetValueNames(Classes::TStrings* Strings);
 
@@ -167,7 +167,7 @@ public:
   virtual void __fastcall WriteDateTime(const UnicodeString Name, TDateTime Value);
   virtual void __fastcall WriteFloat(const UnicodeString Name, double Value);
   virtual void __fastcall WriteStringRaw(const UnicodeString Name, const UnicodeString Value);
-  virtual void __fastcall WriteBinaryData(const UnicodeString Name, const void * Buffer, int Size);
+  virtual void __fastcall WriteBinaryData(const UnicodeString Name, const void * Buffer, size_t Size);
 
   virtual void __fastcall GetValueNames(Classes::TStrings* Strings);
